Move utility highlight modules into "common"
- Move general highlight modules used by other highlight modules into
  .../highlight/common
