+++ src/modules/highlight/common/flow-mutation-observer.d.mts
