+++ src/modules/highlight/engines/paint.mts
@@ -9,8 +9,8 @@ import { getBoxesOwned } from "/dist/modules/highlight/engines/paint/boxes.mjs";
 import type { AbstractTreeCacheEngine } from "/dist/modules/highlight/models/tree-cache.d.mjs";
 import type { AbstractFlowTracker, Flow, Span } from "/dist/modules/highlight/models/tree-cache/flow-tracker.d.mjs";
 import { FlowTracker } from "/dist/modules/highlight/models/tree-cache/flow-tracker.mjs";
-import { highlightTags } from "/dist/modules/highlight/highlight-tags.mjs";
-import * as TermCSS from "/dist/modules/highlight/term-css.mjs";
+import { highlightTags } from "/dist/modules/highlight/common/highlight-tags.mjs";
+import * as TermCSS from "/dist/modules/highlight/common/term-css.mjs";
 import type { MatchTerm, TermTokens, TermPatterns } from "/dist/modules/match-term.mjs";
 import { StyleManager } from "/dist/modules/style-manager.mjs";
 import { HTMLStylesheet } from "/dist/modules/stylesheets/html.mjs";
