+++ src/modules/highlight/tools/term-counters/tree-cache.mts
@@ -5,8 +5,8 @@
  */
 
 import type { AbstractTermCounter } from "/dist/modules/highlight/tools/term-counter.d.mjs";
-import type { BaseFlow } from "/dist/modules/highlight/matcher.mjs";
-import { highlightTags } from "/dist/modules/highlight/highlight-tags.mjs";
+import type { BaseFlow } from "/dist/modules/highlight/common/matcher.mjs";
+import { highlightTags } from "/dist/modules/highlight/common/highlight-tags.mjs";
 import type { MatchTerm } from "/dist/modules/match-term.mjs";
 import type { AllReadonly } from "/dist/modules/common.mjs";
