+++ src/modules/highlight/tools/term-markers/tree-cache.mts
@@ -6,7 +6,7 @@
 
 import type { AbstractTermMarker } from "/dist/modules/highlight/tools/term-marker.d.mjs";
 import { Styles } from "/dist/modules/highlight/tools/term-marker/common.mjs";
-import type { BaseFlow } from "/dist/modules/highlight/matcher.mjs";
+import type { BaseFlow } from "/dist/modules/highlight/common/matcher.mjs";
 import type { MatchTerm, TermTokens } from "/dist/modules/match-term.mjs";
 import { StyleManager } from "/dist/modules/style-manager.mjs";
 import { HTMLStylesheet } from "/dist/modules/stylesheets/html.mjs";
