+++ src/modules/highlight/tools/term-walkers/tree-edit.mts
@@ -6,7 +6,7 @@
 
 import type { AbstractTermWalker } from "/dist/modules/highlight/tools/term-walker.d.mjs";
 import { Styles } from "/dist/modules/highlight/tools/term-walker/common.mjs";
-import { getContainerBlock } from "/dist/modules/highlight/container-blocks.mjs";
+import { getContainerBlock } from "/dist/modules/highlight/common/container-blocks.mjs";
 import { HIGHLIGHT_TAG, HIGHLIGHT_TAG_UPPER } from "/dist/modules/highlight/models/tree-edit/tags.mjs";
 import type { MatchTerm, TermTokens } from "/dist/modules/match-term.mjs";
 import { StyleManager } from "/dist/modules/style-manager.mjs";
